Fix StartGame bug when no game present
For each StartGame request, we expect that it is in response to
a tweet from the SAMQuest bot. If someone tweets @SAMQuest and uses
the #StartGame hashtag which is not a reply to a tweet, this value is
null, and the check in dynamodb fails, causing us to swallow the error.

This change does a simple check to see if the tweet is a reply tweet first,
and sends an error message back if it is not a reply.